+++ src/core/store-id.ts
@@ -3,7 +3,7 @@ import { sitesStore } from '@src/infrastructure/prun-api/data/sites';
 import { warehousesStore } from '@src/infrastructure/prun-api/data/warehouses';
 import { shipsStore } from '@src/infrastructure/prun-api/data/ships';
 
-export function getInvStoreId(invParameter: string) {
+export function getInvStore(invParameter: string) {
   invParameter = invParameter.toLowerCase().trim();
   const stores = storagesStore.all.value ?? [];
   let store = stores.find(x => x.id.startsWith(invParameter));