+++ src/store/user-data-migrations.ts
@@ -1,13 +1,13 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import removeArrayElement from '@src/utils/remove-array-element';
 import { tilesStore } from '@src/infrastructure/prun-api/data/tiles';
-import { getInvStoreId } from '@src/core/store-id';
+import { getInvStore } from '@src/core/store-id';
 
 const migrations: Migration[] = [
   userData => {
     const sorting = {} as Record<string, any>;
     for (const mode of userData.sorting) {
-      const store = getInvStoreId(mode.storeId);
+      const store = getInvStore(mode.storeId);
       if (!store) {
         continue;
       }
@@ -22,7 +22,7 @@ const migrations: Migration[] = [
         continue;
       }
       const storeId = tile.content.substring(3);
-      const store = getInvStoreId(storeId);
+      const store = getInvStore(storeId);
       const state = userData.tileState[tileId];
       if (store) {
         const storeSorting = (sorting[store.id] ??= { modes: [] });